+++ eDnevnik/eDnevnik.Services/Database/KorisnikDetalji.cs
@@ -12,7 +12,7 @@ namespace eDnevnik.Services.Database
         public int? KorisnikID { get; set; }
         public int? GodinaStudija { get; set; }
         public bool? ObnavljaGodinu { get; set; }
-        public double? ProsjecnaOcjena { get; set; }
+        public decimal? ProsjecnaOcjena { get; set; }
         public int? GodinaUpisaID { get; set; }
         public int? UpisanaSkolskaGodinaID { get; set; }
         public virtual SkolskaGodina GodinaUpisa { get; set; } = null!;
